remove logic for archive table migration
