+++ conf/routes
@@ -214,9 +214,5 @@ POST  /support-frontend/default-promos/update              controllers.DefaultPr
 GET   /frontend/super-mode                             controllers.SuperModeController.getSuperModeRows()
 GET   /frontend/epic-article-data                      controllers.SuperModeController.getArticleData()
 
-
-# ----- archived test migration ----- #
-GET  /migrate-archived-tests/:channel                 controllers.ArchivedTestsMigrationController.migrate(channel: String)
-
 # Map static resources from the /public folder to the /assets URL path
 GET  /assets/*file                                  controllers.Assets.at(path="/public", file)
